requirements.txt
grpcio==1.24.3
grpcio-tools==1.24.3
protobuf==3.10.0
six==1.12.0
